% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CC.r
\name{CC}
\alias{CC}
\title{Age-based Catch Curve}
\usage{
CC(CatA, M)
}
\arguments{
\item{CatA}{vector of catch-at-age (in numbers)}

\item{M}{natural mortality rate (assumed constant for age-classes)}
}
\value{
a data.frame with total mortality (Z), fishing mortality (F), and natural mortality (M)
}
\description{
Age-based Catch Curve
}

