% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EqMod.r
\name{EqMod}
\alias{EqMod}
\title{Equilibrium age-structured model}
\usage{
EqMod(M, K, Linf, t0 = 0, Am, As, Fmax, R0 = 1, Steepness = 1)
}
\arguments{
\item{M}{Natural mortality rate (assumed constant for all age-classes)}

\item{K}{von Bertalanffy K parameter}

\item{Linf}{von Bertalanffy Linf parameter}

\item{t0}{von Bertalanffy t0 parameter (assumed 0 for simplicity)}

\item{Am}{age at maturity (knife-edge)}

\item{As}{age at first selectioni (knife-edge)}

\item{Fmax}{Fishing mortality rate for fully selected individuals}

\item{R0}{Initial recruitment (assumed 1)}

\item{Steepness}{Steepness of the Beverton-Holt Stock-Recruitment relationship}
}
\value{
A list of length 2 - first element parameters, second element numbers-at-age
}
\description{
Equilibrium age-structured model
}

