% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_opt.R
\name{m_opt}
\alias{m_opt}
\title{Run logistic regression with variable selection}
\usage{
m_opt(sample_data, p_lim = 0.06, t_sample = 0.8)
}
\arguments{
\item{sample_data}{data frame where 1st column is identifier(key) and 2nd column is target variable should be named flag, data should not have missing values and categorical variables should be one hot encoded}

\item{p_lim}{maximum acceptable p value for variable in logistic regression default 0.06}

\item{t_sample}{ratio of data to be used for training}
}
\value{
data frame (rm_frame) with list of variables removed and reasons, model object base.glm, data frame conf_mat contains confusion matrix and data frame comp_frame comparing test and training set
}
\description{
Returns list of variables removed from model due to variable selection, model object for logistic regression,confusiton matrix and comparison of key accuracy parameters between test and control set, 20% of data passed by default is used for testing
}
