% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_treat.R
\name{var_treat}
\alias{var_treat}
\title{Remove multicollinearity from data}
\usage{
var_treat(data, vif_lim = 10)
}
\arguments{
\item{data}{data frame where 1st column is identifier(key) and 2nd column is target variable should be named flag, data should not have missing values and categorical variables should be one hot encoded}

\item{vif_lim}{maximum acceptable vif value for treating multicollinearity, default set to 10}
}
\value{
data frame (rm_frame) with list of variables removed and data frame (data_treat) containing treated data
}
\description{
Returns list of variables removed due to multicolleniarity and treated data frame
}
